requirements.txt
numpy
pandas
pyyaml
xgboost
scikit-learn
pyarrow
matplotlib
pynvml
